+++ src/modules/facts/actions.js
@@ -1,8 +1,8 @@
-import {selectorSearch} from 'gEngine/factBank'
+import {selectorSearch} from 'gEngine/facts'
 
 export function globalsSearch(selector) {
   return (dispatch, getState) => {
-    const {partial, suggestion} = selectorSearch(selector, getState().factBank.globals)
+    const {partial, suggestion} = selectorSearch(selector, getState().facts.globals)
     dispatch({type: 'SUGGEST_FACT', suggestion: suggestion.replace(partial, '')})
   }
 }
